+++ services/paymaster/src/config_manager/mod.rs
@@ -1,3 +1,2 @@
 pub mod config;
-pub mod fee;
 pub mod load_config;