+++ programs/intent-transfer/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.2 (2026-02-04)
+
+### Fixes
+
+- document imprecision of exec amount computation relative to original logic
+- reduce overflow risk by removing unnecessary normalizations/conversions
+
 ## 0.1.1 (2026-01-08)
 
 ### Fixes